backend/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
composio-langchain>=0.6.15
langchain-community>=0.0.38
langchain-core>=0.1.52

# PDF processing
pypdf>=3.17.1

# API integrations
requests>=2.31.0
aiohttp>=3.9.1

# Utilities
python-dateutil>=2.8.2
typing-extensions>=4.8.0

# Development dependencies
black>=23.12.1  # Code formatting
pylint>=3.0.3   # Code linting
pytest>=7.4.4   # Testing 